+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add proper support for disabled modules
+- Add tree view filtering by severity
 - Fix auth required state of status bar
 
 ## [v1.14.0]